+++ instrumentation-docs/src/main/java/io/opentelemetry/instrumentation/docs/internal/EmittedMetrics.java
@@ -9,18 +9,31 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
- * This class is internal and is hence not for public use. Its APIs are unstable and can change at
- * any time.
+ * Representation of metrics emitted by an instrumentation. Includes context about whether emitted
+ * by default or via a configuration option. This class is internal and is hence not for public use.
+ * Its APIs are unstable and can change at any time.
  */
 public class EmittedMetrics {
+  // Condition in which the metrics are emitted (ex: default, or configuration option names).
+  private String when;
   private List<Metric> metrics;
 
   public EmittedMetrics() {
+    this.when = "";
     this.metrics = new ArrayList<>();
   }
 
-  public EmittedMetrics(List<Metric> metrics) {
+  public EmittedMetrics(String when, List<Metric> metrics) {
+    this.when = "";
     this.metrics = metrics;
+  }
+
+  public String getWhen() {
+    return when;
+  }
+
+  public void setWhen(String when) {
+    this.when = when;
   }
 
   public List<Metric> getMetrics() {
@@ -40,10 +53,14 @@ public class EmittedMetrics {
     private String description;
     private String type;
     private String unit;
-    private List<Attribute> attributes;
+    private List<TelemetryAttribute> attributes;
 
     public Metric(
-        String name, String description, String type, String unit, List<Attribute> attributes) {
+        String name,
+        String description,
+        String type,
+        String unit,
+        List<TelemetryAttribute> attributes) {
       this.name = name;
       this.description = description;
       this.type = type;
@@ -91,47 +108,12 @@ public class EmittedMetrics {
       this.unit = unit;
     }
 
-    public List<Attribute> getAttributes() {
+    public List<TelemetryAttribute> getAttributes() {
       return attributes;
     }
 
-    public void setAttributes(List<Attribute> attributes) {
+    public void setAttributes(List<TelemetryAttribute> attributes) {
       this.attributes = attributes;
-    }
-  }
-
-  /**
-   * This class is internal and is hence not for public use. Its APIs are unstable and can change at
-   * any time.
-   */
-  public static class Attribute {
-    private String name;
-    private String type;
-
-    public Attribute() {
-      this.name = "";
-      this.type = "";
-    }
-
-    public Attribute(String name, String type) {
-      this.name = name;
-      this.type = type;
-    }
-
-    public String getName() {
-      return name;
-    }
-
-    public void setName(String name) {
-      this.name = name;
-    }
-
-    public String getType() {
-      return type;
-    }
-
-    public void setType(String type) {
-      this.type = type;
     }
   }
 }