+++ instrumentation/aws-sdk/aws-sdk-2.2/testing/src/main/java/io/opentelemetry/instrumentation/awssdk/v2_2/AbstractAws2ClientTest.java
@@ -14,6 +14,9 @@ import static io.opentelemetry.semconv.ServerAttributes.SERVER_ADDRESS;
 import static io.opentelemetry.semconv.ServerAttributes.SERVER_PORT;
 import static io.opentelemetry.semconv.UrlAttributes.URL_FULL;
 import static io.opentelemetry.semconv.incubating.AwsIncubatingAttributes.AWS_REQUEST_ID;
+import static io.opentelemetry.semconv.incubating.AwsIncubatingAttributes.AWS_SECRETSMANAGER_SECRET_ARN;
+import static io.opentelemetry.semconv.incubating.AwsIncubatingAttributes.AWS_STEP_FUNCTIONS_ACTIVITY_ARN;
+import static io.opentelemetry.semconv.incubating.AwsIncubatingAttributes.AWS_STEP_FUNCTIONS_STATE_MACHINE_ARN;
 import static io.opentelemetry.semconv.incubating.MessagingIncubatingAttributes.MESSAGING_DESTINATION_NAME;
 import static io.opentelemetry.semconv.incubating.MessagingIncubatingAttributes.MESSAGING_MESSAGE_ID;
 import static io.opentelemetry.semconv.incubating.MessagingIncubatingAttributes.MESSAGING_OPERATION;
@@ -79,6 +82,18 @@ import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.S3ClientBuilder;
 import software.amazon.awssdk.services.s3.model.CreateBucketRequest;
 import software.amazon.awssdk.services.s3.model.GetObjectRequest;
+import software.amazon.awssdk.services.secretsmanager.SecretsManagerAsyncClient;
+import software.amazon.awssdk.services.secretsmanager.SecretsManagerAsyncClientBuilder;
+import software.amazon.awssdk.services.secretsmanager.SecretsManagerClient;
+import software.amazon.awssdk.services.secretsmanager.SecretsManagerClientBuilder;
+import software.amazon.awssdk.services.secretsmanager.model.GetSecretValueRequest;
+import software.amazon.awssdk.services.secretsmanager.model.GetSecretValueResponse;
+import software.amazon.awssdk.services.sfn.SfnAsyncClient;
+import software.amazon.awssdk.services.sfn.SfnAsyncClientBuilder;
+import software.amazon.awssdk.services.sfn.SfnClient;
+import software.amazon.awssdk.services.sfn.SfnClientBuilder;
+import software.amazon.awssdk.services.sfn.model.DescribeActivityRequest;
+import software.amazon.awssdk.services.sfn.model.DescribeStateMachineRequest;
 import software.amazon.awssdk.services.sns.SnsAsyncClient;
 import software.amazon.awssdk.services.sns.SnsAsyncClientBuilder;
 import software.amazon.awssdk.services.sns.SnsClient;
@@ -111,6 +126,18 @@ public abstract class AbstractAws2ClientTest extends AbstractAws2ClientCoreTest 
       "<DeleteOptionGroupResponse xmlns=\"http://rds.amazonaws.com/doc/2014-09-01/\">"
           + " <ResponseMetadata><RequestId>0ac9cda2-bbf4-11d3-f92b-31fa5e8dbc99</RequestId></ResponseMetadata>"
           + "</DeleteOptionGroupResponse>";
+
+  private static final String secretsManagerBodyContent =
+      "{"
+          + "    \"ARN\": \"arn:aws:secretsmanager:us-east-1:123456789012:secret:MySecretFromCLI-sNkBwD\","
+          + "    \"Name\": \"MySecretFromCLI\","
+          + "    \"VersionId\": \"9959b95b-1234-5678-a19b-a4b0315ca5aa\","
+          + "    \"SecretString\": \"super-secret-value\","
+          + "    \"VersionStages\": ["
+          + "        \"AWSCURRENT\""
+          + "    ],"
+          + "    \"CreatedDate\": \"1.523477145713E9\""
+          + "}";
 
   private static void assumeSupportedConfig(String operation) {
     Assumptions.assumeFalse(
@@ -212,6 +239,21 @@ public abstract class AbstractAws2ClientTest extends AbstractAws2ClientCoreTest 
                   equalTo(MESSAGING_OPERATION, "publish"),
                   satisfies(MESSAGING_MESSAGE_ID, val -> val.isInstanceOf(String.class)),
                   equalTo(MESSAGING_SYSTEM, AWS_SQS))));
+    }
+
+    if (service.equals("Sfn")) {
+      if (operation.equals("DescribeStateMachine")) {
+        attributes.add(equalTo(AWS_STEP_FUNCTIONS_STATE_MACHINE_ARN, "state:machine:arn:foo"));
+      } else if (operation.equals("DescribeActivity")) {
+        attributes.add(equalTo(AWS_STEP_FUNCTIONS_ACTIVITY_ARN, "activity:arn:foo"));
+      }
+    }
+
+    if (service.equals("SecretsManager")) {
+      attributes.add(
+          equalTo(
+              AWS_SECRETSMANAGER_SECRET_ARN,
+              "arn:aws:secretsmanager:us-east-1:123456789012:secret:MySecretFromCLI-sNkBwD"));
     }
 
     String evaluatedOperation;
@@ -721,5 +763,125 @@ public abstract class AbstractAws2ClientTest extends AbstractAws2ClientCoreTest 
         .isInstanceOf(SdkException.class);
 
     assertThat(Context.current()).isEqualTo(Context.root());
+  }
+
+  private static Stream<Arguments> provideStepFunctionsArguments() {
+    return Stream.of(
+        Arguments.of(
+            (Function<SfnClient, Object>)
+                c ->
+                    c.describeStateMachine(
+                        DescribeStateMachineRequest.builder()
+                            .stateMachineArn("state:machine:arn:foo")
+                            .build()),
+            "DescribeStateMachine",
+            "POST",
+            "UNKNOWN"),
+        Arguments.of(
+            (Function<SfnClient, Object>)
+                c ->
+                    c.describeActivity(
+                        DescribeActivityRequest.builder().activityArn("activity:arn:foo").build()),
+            "DescribeActivity",
+            "POST",
+            "UNKNOWN"));
+  }
+
+  @ParameterizedTest
+  @MethodSource("provideStepFunctionsArguments")
+  void testSfnSendOperationRequestWithBuilder(
+      Function<SfnClient, Object> call, String operation, String method, String requestId) {
+    SfnClientBuilder builder = SfnClient.builder();
+    configureSdkClient(builder);
+    SfnClient client =
+        builder
+            .endpointOverride(clientUri)
+            .region(Region.AP_NORTHEAST_1)
+            .credentialsProvider(CREDENTIALS_PROVIDER)
+            .build();
+
+    server.enqueue(HttpResponse.of(HttpStatus.OK, MediaType.PLAIN_TEXT_UTF_8, ""));
+    Object response = call.apply(client);
+    assertThat(response.getClass().getSimpleName())
+        .satisfiesAnyOf(
+            v ->
+                assertThat(response)
+                    .isInstanceOf(
+                        software.amazon.awssdk.services.sfn.model.DescribeActivityResponse.class),
+            v ->
+                assertThat(response)
+                    .isInstanceOf(
+                        software.amazon.awssdk.services.sfn.model.DescribeStateMachineResponse
+                            .class));
+    clientAssertions("Sfn", operation, method, response, requestId);
+  }
+
+  @ParameterizedTest
+  @MethodSource("provideStepFunctionsArguments")
+  void testSfnAsyncSendOperationRequestWithBuilder(
+      Function<SfnClient, Object> call, String operation, String method, String requestId) {
+    SfnAsyncClientBuilder builder = SfnAsyncClient.builder();
+    configureSdkClient(builder);
+    SfnAsyncClient client =
+        builder
+            .endpointOverride(clientUri)
+            .region(Region.AP_NORTHEAST_1)
+            .credentialsProvider(CREDENTIALS_PROVIDER)
+            .build();
+
+    server.enqueue(HttpResponse.of(HttpStatus.OK, MediaType.PLAIN_TEXT_UTF_8, ""));
+    Object response = call.apply(wrapClient(SfnClient.class, SfnAsyncClient.class, client));
+    assertThat(response.getClass().getSimpleName())
+        .satisfiesAnyOf(
+            v ->
+                assertThat(response)
+                    .isInstanceOf(
+                        software.amazon.awssdk.services.sfn.model.DescribeActivityResponse.class),
+            v ->
+                assertThat(response)
+                    .isInstanceOf(
+                        software.amazon.awssdk.services.sfn.model.DescribeStateMachineResponse
+                            .class));
+    clientAssertions("Sfn", operation, method, response, requestId);
+  }
+
+  @Test
+  void testSecretsManagerSendOperationRequestWithBuilder() {
+    SecretsManagerClientBuilder builder = SecretsManagerClient.builder();
+    configureSdkClient(builder);
+    SecretsManagerClient client =
+        builder
+            .endpointOverride(clientUri)
+            .region(Region.AP_NORTHEAST_1)
+            .credentialsProvider(CREDENTIALS_PROVIDER)
+            .build();
+
+    server.enqueue(
+        HttpResponse.of(HttpStatus.OK, MediaType.PLAIN_TEXT_UTF_8, secretsManagerBodyContent));
+    Object response =
+        client.getSecretValue(GetSecretValueRequest.builder().secretId("MySecretFromCLI").build());
+    assertThat(response.getClass().getSimpleName())
+        .satisfies(
+            v -> assertThat(v).isEqualTo("GetSecretValueResponse"),
+            v -> assertThat(response).isInstanceOf(GetSecretValueResponse.class));
+    clientAssertions("SecretsManager", "GetSecretValue", "POST", response, "UNKNOWN");
+  }
+
+  @Test
+  void testSecretsManagerAsyncSendOperationRequestWithBuilder() {
+    SecretsManagerAsyncClientBuilder builder = SecretsManagerAsyncClient.builder();
+    configureSdkClient(builder);
+    SecretsManagerAsyncClient client =
+        builder
+            .endpointOverride(clientUri)
+            .region(Region.AP_NORTHEAST_1)
+            .credentialsProvider(CREDENTIALS_PROVIDER)
+            .build();
+
+    server.enqueue(
+        HttpResponse.of(HttpStatus.OK, MediaType.PLAIN_TEXT_UTF_8, secretsManagerBodyContent));
+    Object response =
+        client.getSecretValue(GetSecretValueRequest.builder().secretId("MySecretFromCLI").build());
+    clientAssertions("SecretsManager", "GetSecretValue", "POST", response, "UNKNOWN");
   }
 }