+++ instrumentation/spring/spring-boot-autoconfigure/src/main/java/io/opentelemetry/instrumentation/spring/autoconfigure/OpenTelemetryAutoConfiguration.java
@@ -41,6 +41,7 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.DependsOn;
 import org.springframework.core.env.Environment;
 
 /**
@@ -62,17 +63,25 @@ public class OpenTelemetryAutoConfiguration {
 
   public OpenTelemetryAutoConfiguration() {}
 
+  @Bean
+  @ConfigurationPropertiesBinding
+  OtelMapConverter otelMapConverter() {
+    // This is needed for otlp exporter headers and OtelResourceProperties.
+
+    // We need this converter, even if the SDK is disabled,
+    // because the properties are parsed before the SDK is disabled.
+
+    // We also need this converter if the OpenTelemetry bean is user supplied,
+    // because the environment variables may still contain a value that needs to be converted,
+    // even if the SDK is disabled (and the value thus ignored).
+    return new OtelMapConverter();
+  }
+
   @Configuration
   @Conditional(SdkEnabled.class)
+  @DependsOn("otelMapConverter")
   @ConditionalOnMissingBean(OpenTelemetry.class)
   static class OpenTelemetrySdkConfig {
-
-    @Bean
-    @ConfigurationPropertiesBinding
-    public OtelMapConverter otelMapConverter() {
-      // needed for otlp exporter headers and OtelResourceProperties
-      return new OtelMapConverter();
-    }
 
     @Bean
     public OpenTelemetrySdkComponentLoader openTelemetrySdkComponentLoader(
@@ -139,21 +148,10 @@ public class OpenTelemetryAutoConfiguration {
   }
 
   @Configuration
+  @DependsOn("otelMapConverter")
   @ConditionalOnMissingBean(OpenTelemetry.class)
   @ConditionalOnProperty(name = "otel.sdk.disabled", havingValue = "true")
   static class DisabledOpenTelemetrySdkConfig {
-
-    @Bean
-    @ConfigurationPropertiesBinding
-    // Duplicated in OpenTelemetrySdkConfig and DisabledOpenTelemetrySdkConfig to not expose the
-    // converter in the public API
-    public OtelMapConverter otelMapConverter() {
-      // needed for otlp exporter headers and OtelResourceProperties
-      // we need this converter, even if the SDK is disabled,
-      // because the properties are parsed before the SDK is disabled
-      return new OtelMapConverter();
-    }
-
     @Bean
     public OpenTelemetry openTelemetry() {
       logger.info("OpenTelemetry Spring Boot starter has been disabled");